five8

Function encode_64

pub fn encode_64(bytes: &[u8; 64], out: &mut [u8; 88]) -> u8

Encode a 64-byte array.

Mutates the provided out array and returns a u8 len which indicates how many bytes of the out array were actually written to. The remaining bytes are unchanged, hence the result we care about after calling the function is out[..len as usize].

Examples

let mut buf = [0u8; 88];
let bytes = &[
    0, 0, 10, 85, 198, 191, 71, 18, 5, 54, 6, 255, 181, 32, 227, 150, 208, 3, 157, 135, 222,
    67, 50, 23, 237, 51, 240, 123, 34, 148, 111, 84, 98, 162, 236, 133, 31, 93, 185, 142, 108,
    41, 191, 1, 138, 6, 192, 0, 46, 93, 25, 65, 243, 223, 225, 225, 85, 55, 82, 251, 109, 132,
    165, 2,
];
let len = five8::encode_64(bytes, &mut buf);
assert_eq!(
    &buf[..len as usize],
    [
        49, 49, 99, 103, 84, 72, 52, 68, 53, 101, 56, 83, 51, 115, 110, 68, 52, 52, 52, 87, 98,
        98, 71, 114, 107, 101, 112, 106, 84, 118, 87, 77, 106, 50, 106, 107, 109, 67, 71, 74, 116,
        103, 110, 51, 72, 55, 113, 114, 80, 98, 49, 66, 110, 119, 97, 112, 120, 112, 98, 71, 100,
        82, 116, 72, 81, 104, 57, 116, 57, 87, 98, 110, 57, 116, 54, 90, 68, 71, 72, 122, 87, 112,
        76, 52, 100, 102
    ]
);
assert_eq!(len, 86);